AlmacenarEmbedding/requirements.txt
flask
requests
numpy
firebase-admin
functions-framework
google-auth